Add FIRST/LAST constants, move client code
